requirements.txt
gradio
opencv-python
pillow
numpy
